Add base class in preparation for splitting up RoundtripTest
